remove redis

+++ src/operation/blog.py
@@ -1,5 +1,6 @@
 import base64
 import hashlib
+import logging
 import re
 import asyncio
 import datetime
@@ -14,7 +15,7 @@ from multiprocessing import Process
 from pydantic import BaseModel, validator
 from xpinyin import Pinyin
 from src import utils
-from src.db.client.my_redis import GlobalLock, redis_client
+from src.db.client.my_redis import REDIS_URL, redis_client
 from src.framework.error import ErrorWithPrompt
 from src.framework.config import BLOG_ROOT, STORAGE_ROOT
 from src.operation import data_io
@@ -276,6 +277,7 @@ class BlogBuilder:
 
         # 存放生成好的静态资源的额 root 目录
         write_root = os.path.join(BLOG_ROOT, user, service)
+        logging.debug(f"start generate blog html files for email: {email}, write to ->\n\t{write_root}")
         try:
             shutil.rmtree(write_root)
         except:  # noqa
@@ -313,10 +315,11 @@ async def fresh_blog(email: str):
     BLOG_ROOT/{email_user}/{service}/version.txt  此文件夹下的文件名代表刷新 blog 时的时间戳
     BLOG_ROOT/{email_user}/{service}/index.html  此文件夹下
     """
-    uni_key = f"genb:{email}"
-    lock = await redis_client.set_if_not_exists(uni_key, value="123", timeout="3600")
-    if not lock:
-        raise ErrorWithPrompt("Blog正在生成中，请稍后再试")
+    if REDIS_URL:
+        uni_key = f"genb:{email}"
+        lock = await redis_client.set_if_not_exists(uni_key, value="123", timeout="3600")
+        if not lock:
+            raise ErrorWithPrompt("Blog正在生成中，请稍后再试")
 
     async def async_wrapper():
         p = Process(target=gen_wrapper, args=(email, ))
@@ -327,6 +330,7 @@ async def fresh_blog(email: str):
         while p.is_alive():
             await asyncio.sleep(3)
 
-        await redis_client.delete(uni_key)
+        if REDIS_URL:
+            await redis_client.delete(uni_key)
 
     asyncio.create_task(async_wrapper())
